Renamed connection classes to better reflect their functionality
